"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ cspp2-assignment/m2/Assignment-3/Solution.java
@@ -6,18 +6,18 @@ public class Solution {
 	*/
 	public static void main(String[] args) {
 		Scanner s=new Scanner(System.in);      
-        int base = s.nextInt();
-        int exponent = s.nextInt();
-        int result=power(base,exponent);
+        long base = s.nextLong();
+        long exponent = s.nextLong();
+        long result=power(base,exponent);
         System.out.println(result);
 	}
 	/*
 	Need to write the power function and print the output.
 	*/
-	public static int power(int a, int b) {
+	public static long power(long a, long b) {
 		if (b == 0)
 			return 1;
-		int result = a * power(a, b-1);
+		long result = a * power(a, b-1);
 		return result;
 
 	}
